refactor(js): extract `svgMe` function to `utils/svg-me.js` module
Steps:
- Create `assets/js/utils/svg-me.js` with exported `svgMe` function
- Remove `svgMe` function from `scripts.js`
- Add `import { svgMe }` to `scripts.js`

Reason:
- Part of the modularization effort to split `scripts.js` into separate ES modules

References:
- Issue #104
